modal modalの表示するコードの修正
