+++ resources/js/top/index.js
@@ -1,42 +1,6 @@
-import * as Vue from "vue";
-import { ref } from "vue";
-import SampleComponent from "./SampleComponent.vue";
+import { createApp } from 'vue';
+import Modal from './Modal.vue';
 
-const application = {
-    setup() {
-
-        const page = ref(1);
-
-        const name = ref("");
-        const validateResult = ref("");
-
-        const title = ref("vue test title.");
-
-        let buttonClick = function() {
-            title.value = "value update.";
-        };
-
-        let validate = function(){
-            let isKana = name.value.match(/^[あーんー ]*$/);
-            validateResult.value = isKana ? "正常" : "ひらがな以外が入力されています。";
-        }
-        return {
-            title,
-            buttonClick,
-
-            name,
-            validateResult,
-            validate,
-
-        };
-    },
-
-    components : {
-        "sample-component" : SampleComponent,
-    }
-};
-
-try {
-    let mainElement = document.getElementsByTagName("main")[0];
-    Vue.createApp(application).mount(mainElement);
-} catch (e) {}
+const app = createApp({});
+app.component('modal', Modal);
+app.mount('#app');
